nira_Q&A/requirement.txt
fastapi
uvicorn
aiofiles
aiohttp
httpx
python-dotenv
watchdog
nltk
sentence-transformers
pinecone-client
ollama
huggingface_hub